feat(stylist): add stylist integration to a crate
