+++ crates/stellation-stylist/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "stellation-stylist"
+version = "0.2.0"
+edition = "2021"
+rust-version = "1.66"
+repository = "https://github.com/futursolo/stellation"
+authors = [
+    "Kaede Hoshiakwa <futursolo@icloud.com>",
+]
+description = "The framework experience for Yew."
+keywords = ["web", "wasm", "yew", "framework", "ssr"]
+categories = ["wasm", "web-programming"]
+readme = "../../README.md"
+homepage = "https://github.com/futursolo/stellation"
+license = "MIT OR Apache-2.0"
+
+[dependencies]
+stylist = { version = "0.12.1", features = ["yew_integration"] }
+yew = "0.20"
+stellation-backend = { version = "0.2.0", path = "../stellation-backend", optional = true }
+
+[features]
+frontend = ["stylist/ssr"]
+backend = ["stylist/hydration", "dep:stellation-backend"]
+
+[package.metadata.docs.rs]
+all-features = true
+rustdoc-args = ["--cfg", "documenting"]
