+++ examples/fullstack/server/Cargo.toml
@@ -11,12 +11,12 @@ anyhow = "1"
 tokio = { version = "1.29.1", features = ["full"] }
 tracing = { version = "0.1.37" }
 yew = "0.20.0"
-stylist = { version = "0.12.1", features = ["yew_integration", "ssr"] }
 
 # Stellation Components
 stellation-backend = { version = "0.2.0", path = "../../../crates/stellation-backend" }
 stellation-backend-tower = { version = "0.2.0", path = "../../../crates/stellation-backend-tower" }
 stellation-backend-cli = { version = "0.2.0", path = "../../../crates/stellation-backend-cli" }
+stellation-stylist = { version = "0.2.0", path = "../../../crates/stellation-stylist", features = ["backend"] }
 
 # Example Workspace
 example-fullstack-view = { path = "../view" }
